Switch to ES6 modules & bundle/minify GeoJSON
