+++ .parcelrc
@@ -1,7 +1,7 @@
 {
   "extends": "@parcel/config-default",
   "transformers": {
-    "*.geojson": ["@parcel/transformer-raw"],
+    "*.geojson": ["@parcel/transformer-json"],
     "*.js": ["@parcel/transformer-js"]
   }
 }
